+++ NATS.Jwt/Models/NatsAuthorizationResponseClaims.cs
@@ -1,30 +1,20 @@
+// Copyright (c) The NATS Authors.
+// Licensed under the Apache License, Version 2.0.
+
 using System.Text.Json.Serialization;
 
 namespace NATS.Jwt.Models;
 
+/// <summary>
+/// Represents the claims data for a NATS authorization response.
+/// </summary>
 public record NatsAuthorizationResponseClaims : JwtClaimsData
 {
+    /// <summary>
+    /// Gets or sets the value representing an authorization response for NATS.
+    /// </summary>
     [JsonPropertyName("nats")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     [JsonPropertyOrder(1)]
     public NatsAuthorizationResponse AuthorizationResponse { get; set; } = new();
-}
-
-public record NatsAuthorizationResponse : NatsGenericFields
-{
-    [JsonPropertyName("jwt")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public string Jwt { get; set; }
-
-    [JsonPropertyName("error")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public string Error { get; set; }
-
-    /// <summary>
-    /// IssuerAccount stores the public key for the account the issuer represents.
-    /// When set, the claim was issued by a signing key.
-    /// </summary>
-    [JsonPropertyName("issuer_account")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public string IssuerAccount { get; set; }
 }
