+++ NATS.Jwt/Models/NatsOperator.cs
@@ -1,12 +1,18 @@
+// Copyright (c) The NATS Authors.
+// Licensed under the Apache License, Version 2.0.
+
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
 namespace NATS.Jwt.Models;
 
+/// <summary>
+/// Represents a NATS operator.
+/// </summary>
 public record NatsOperator : NatsGenericFields
 {
     /// <summary>
-    /// List of other operator NKeys that can be used to sign on behalf of the main
+    /// Gets or sets a list of other operator NKeys that can be used to sign on behalf of the main
     /// operator identity.
     /// </summary>
     [JsonPropertyName("signing_keys")]
@@ -14,7 +20,7 @@ public record NatsOperator : NatsGenericFields
     public List<string> SigningKeys { get; set; }
 
     /// <summary>
-    /// AccountServerURL is a partial URL like "https://host.domain.org:<c>port</c>/jwt/v1"
+    /// Gets or sets account server URL which is a partial URL like "https://host.domain.org:<c>port</c>/jwt/v1"
     /// tools will use the prefix and build queries by appending /accounts/<c>account_id</c>
     /// or /operator to the path provided. Note this assumes that the account server
     /// can handle requests in a nats-account-server compatible way. See
@@ -25,7 +31,7 @@ public record NatsOperator : NatsGenericFields
     public string AccountServerUrl { get; set; }
 
     /// <summary>
-    /// A list of NATS urls (tls://host:port) where tools can connect to the server
+    /// Gets or sets a list of NATS urls (tls://host:port) where tools can connect to the server
     /// using proper credentials.
     /// </summary>
     [JsonPropertyName("operator_service_urls")]
@@ -33,21 +39,21 @@ public record NatsOperator : NatsGenericFields
     public List<string> OperatorServiceUrLs { get; set; }
 
     /// <summary>
-    /// Identity of the system account
+    /// Gets or sets an identity of the system account.
     /// </summary>
     [JsonPropertyName("system_account")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public string SystemAccount { get; set; }
 
     /// <summary>
-    /// Min Server version
+    /// Gets or sets min Server version.
     /// </summary>
     [JsonPropertyName("assert_server_version")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public string AssertServerVersion { get; set; }
 
     /// <summary>
-    /// Signing of subordinate objects will require signing keys
+    /// Gets or sets a value indicating whether signing of subordinate objects will require signing keys.
     /// </summary>
     [JsonPropertyName("strict_signing_key_usage")]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
